Add helpers for writing parser fscheck properties
I want to be able to write fscheck properties that return true if a parser succeeds as expected
(or, false if it fails). The `testParserSucceedsWith` function lets me do that...

> testParserSucceedsWith pfloat ("3.14", 3.14);;
val it : bool = true
> testParserSucceedsWith pfloat ("3.14", 42.0);;
val it : bool = false
> testParserSucceedsWith pfloat ("xxx", 42.0);;
val it : bool = false

Conversely, `testParserFails` helps to write fscheck properties to return true if an invalid
parse fails as expected...

> testParserFails pfloat "xxx";;
val it : bool = true
> testParserFails pfloat "3.14";;
val it : bool = false

Note that in the `testParser` function that they both use, `run parser input` is a
convenient abbreviation for `runParserOnString parser () "" input`. I think I'll have to
switch to the latter if I introduce a user state to the parser later.

See..
http://www.quanttec.com/fparsec/reference/charparsers.html#members.run
http://www.quanttec.com/fparsec/reference/charparsers.html#members.runParserOnString

Details of the `ParserResult` union type returned by the `run` function can be found
here - http://www.quanttec.com/fparsec/reference/charparsers.html#members.ParserResult -
which should make the match cases in the `testParser` function clear.