Implemented the isotope parser
The simplest primitive of the SMILES parser...

Excerpt from the OpenSMILES specification:

> An isotope is interpreted as a number,so that [2H], [02H] and [002H] all mean
deuterium. If the isotope field is not specified then the atom is assumed to have
the naturally-occurring isotopic ratios. The isotope value 0 also indicates an
isotope of zero, that is [0S] is not the same as [S]. There is no requirement
that the isotope is a genuine isotope of the element. Thus, [36Cl] is allowed
even though 35Cl and 37Cl are the actual known stable isotopes of chlorine.
A general-purpose SMILES parser must accept at least three digits for the isotope
and values from 0 to 999

The isotope parser doesn't address the optional nature of the isotope field
described above - that will be dealt with at the point this parser is applied.
It simply uses the FParsec primitive parser `puint32` to parses any string that
can be interpreted as a non-negative natural number below 1000.

Note: The FParsec documentation advices against using the computation expression
"monadic syntax" for performance reasons...
...http://www.quanttec.com/fparsec/users-guide/where-is-the-monad.html

I find the "monadic syntax" more readable and I don't expect SMILES strings to
be long enough that I need to worry about performance. So, I've started off using
it here. I may change this later when I become more acquainted with the FParsec
syntax.