Generators for testing element symbols parser
This is the list of element symbols from the OpenSMILES specification. It looks like it's
just the 114 confirmed elements (without "ununtrium" etc).

`elementSymbolGenerator` will randomly choose any one of the symbols, and
`notAnElementSymbolGenerator` generates random strings which are guaranteed to
not start with an element symbol.

Note `elementSymbolGenerator` returns the `symbol` twice in a tuple so it can be
used with `testParserSucceedsWith` which expects a parser and a 2-element tuple
where the first element is the string to be parsed and the second element is the
expected result from the parser. In this case, the parser is expected to simply
return the symbol.