Implemented the elementSymbol parser
Parses any of the 114 confirmed element symbols as a string. (According to the OpenSMILES
specification, this is all it needs to support - I suppose it wouldn't be too much effort to add
ununtrium,  etc, as well though.

I've stuck the single character symbols at the end to avoid clashes (e.g., "Bk" getting parsed
as "B"). The fparsec `choice` parser tries the given parsers from left to right until one succeeds.
This does mean that, given a single letter symbol, it has to try all of the two letter symbols first
so I may have to look at this again if it causes a problem.

Having said that though, H atoms are normally omitted as 'impicit' in a SMILES string and most
of the other single-letter elements are part of the special "organic subset" of atoms (B, C, N, O, P,
S, F, Cl, Br, I) which can be written in a SMILES string without the brackets.