Generalise the option generators
Extracted (and generalised) the generators for choosing an option from a list of strings, and for generating
a random guaranteed to not start with an element symbol. (The latter can be used to test some cases
where a parser is expected to fail.)

They'll going to be useful for testing any parser that choosed from a list of possibilities.

Note the `parserInputOutputChoice` returns the choice twice in a tuple so it can be
used with `testParserSucceedsWith` which expects a parser and a 2-element tuple
where the first element is the string to be parsed and the second element is the
expected result from the parser.