Improved the ParserTestHelper functions
`testParserSucceedsWith` takes a tuple parameter where the first item is the
input string to parse and the second item is the expected result of the parser.

Several parsers simply need to parse a particular string and return that
as the result (otherwise fail). So to avoid having to pass in a tuple like
`(input, input)` I've added testParserSucceedsWithInput` to use for testing
such parsers.

This also means there's no longer a need for the `parserInputChoice` generator
(which admittedly felt a little awkard when I wrote it!). Instead I can use
`choiceGenerator` directly with `testParserSucceedsWithInput`