Implemented a chirality parser
It makes use of the `prefixSuffix` & `pint32Range` parsers as the chirality is represented
in a SMILES string as a prefix (e.g., "@TH", "@AL", etc) followed by an integer in a
particular range. (The exceptions being the shorthand representations "@" and "@@"
for anti-clockwise and clockwise chirality respectively.)

I'm not attaching any meaning to the chirality just yet, so for now the result of the
chiral parser is just the string from the SMILES input representing the chirality.

The `chiralityGenerator` in the tests isn't actually used yet, but I'm expecting it to be
useful later for building up random SMILES strings for other tests.