Implemented a H count parser
If H is speciﬁed without a number, it is identical to H1.  Otherwise, it's speciﬁed as Hn where
n is a number such as H3.

This parser is not responsible for detecting an H count with an invalid n. It doesn't know if
the character after the H is meant to be a count or if it's an implied H1 and the next character
is the start of the next part of the SMILES string. That'll be up to the parser that uses this parser
to work out!

(That's why `invalidHcountFails` just tests that it fails with any string that doesn't start with 'H'
...there's no test for a string that starts with 'H' followed by anything that isn't a number.)

Other related things that this parser doesn't deal with (so will need to be included elsewhere):

+ If no Hn is speciﬁed, it is identical to H0.
+ A hydrogen atom cannot have a hydrogen count, for example, [HH1] is illegal.