Implement a charge parser
Charge is speciﬁed by a +n or -n where n is a number; if the number is missing, it means either
+1 or -1 as appropriate. An implementation is required to accept charges in the range -15 to +15.

First it applies a parser (sign) which turns a '+' or a '-' into a function which does nothing
or a function which changes the sign of a number (respectively). Next it parses a positive
integer (0-15) - defaulting to 1 if the parser fails. Finally it applies the function returned
from the first parser to the number returned from the second.

I toyed with writing this without using the monadic syntax as the documentation recommends
(http://www.quanttec.com/fparsec/users-guide/where-is-the-monad.html)
so the main body of the charge parser would look something like this...

    sign .>>. ((pint32Range 0 15) <|>% 1) |>> (fun (f, v) -> f v)

...which does make sense to me now but I still think the monadic version is much more
readable and I can't be sure that if I come back to this in the future it'll still make sense
to me!