Introduce a padded number generator
Some parts of the SMILES specification that accept digits specify that the digits should be
interpreted as a number regardless of leading zeros. E.g., "An isotope is interpreted as a number,
so that [2H], [02H] and [002H] all mean deuterium."

The atom class specification has something similar, so before making the primitive for atomClass
I've extracted out from the `validIsotopeGenerator` a generator that generates a number with
optional leading zeros.

For convenience it generates a tuple where the first element is the number as a string with optional
leading zeros, and the second element is the original number.