Make sure the test modules are in the FMol.Tests namespace
I'm still trying to work out best practices for organising code into modules
and namespaces in an F# project - especially if the default "Add item" doesn't
automatically add a namespace to the module. It's easy to inadvertently add
things to the wrong namespace (which is what happened here).

For now, at least, I'm going to try and be consistent with the system I started
out with!