Handle an atom with an implicit hydrogen count
This is so I can make an atom from the `aliphatic_organic`, `aromatic_organic`, and `unknown`
parsers. These atoms can be written using only the atomic symbol (that is, without the
square brackets, H-count, etc.).

These atoms have "implicit hydrogens" which are added such that the valence of the atom
is in the lowest normal state for that element. So, creating an atom with 0 hCount is clearly
wrong and turning the hCount into an `int option` doesn't quite specify the intent as well
as a discrimant union of either an int value or "Implicit".