Implement the atom parser
bracket_atom | aliphatic_organic | aromatic_organic | ’*’

The atom parser needs to generate an `Atom`. For the organic & unknown parsers, the
generated atom will:

- have "implicit hydrogens" (i.e., they're added such that valence of the atom
  is in the lowest normal state for that element - but that'll most likely happen
  as a post-process step much later!)
- have zero charge
- have no isotopic specification
- have no chiral specification

The `makeAtom` function takes the symbol parsed from an organic or unknown parser
and turns it into an `Atom` with the above characteristics. The atom parser is
then either the result of this or a `bracketAtom`.