Fix invalidIsotopeGenerator and invalidAtomClassGenerator
...so that they can't generate strings starting with "+" or "-" followed by one
or more digits.

The openSMILES specification defines `isotope` as a NUMBER and `atom class` as
a ":" followed by a NUMBER. While I'm assuming that NUMBER implies a whole
number, it's unclear whether it allows negative numbers or positive numbers with
an explicit "+".

The implementations of the `isotope` and `atomClass` parsers will successfully
parse strings like "+42" or "-36" and ":+42" and ":-36" (respectively) so (unless
I find out otherwise) I'm going to allow these as valid and therefore disallow
them from the invalid tests.